Only cleanup at very end once, or cleanup if cancellation

+++ .gitlab/utils/cancel-flux.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 set -x
+
+NO_CLEAN=false
+if [[ "$1" == "--no-clean" ]]; then
+    NO_CLEAN=true
+fi
+
 export URI=$(flux jobs -o "{id} {name}" | grep ${ALLOC_NAME}${GPUMODE} | awk '{print $1}')
 ([[ -n "${URI}" ]] && flux job kill ${URI} || exit 0)
-echo "Removing $CUSTOM_CI_BUILDS_DIR"
-rm -rf $CUSTOM_CI_BUILDS_DIR
+
+if ! $NO_CLEAN; then
+    rm -rf $CUSTOM_CI_BUILDS_DIR
+fi
