+++ .gitlab/utils/cancel-slurm.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 set -x
+
+NO_CLEAN=false
+if [[ "$1" == "--no-clean" ]]; then
+    NO_CLEAN=true
+fi
+
 export JOBID=$(squeue -h --name=${ALLOC_NAME} --format=%A)
 ([[ -n "${JOBID}" ]] && scancel ${JOBID} || exit 0)
-echo "Removing $CUSTOM_CI_BUILDS_DIR"
-rm -rf $CUSTOM_CI_BUILDS_DIR
+
+if ! $NO_CLEAN; then
+    rm -rf $CUSTOM_CI_BUILDS_DIR
+fi
